refactor: move unused functions causing cyclic import moved
