+++ src/ai/backend/manager/services/session/types.py
@@ -1,10 +1,38 @@
+from __future__ import annotations
+
 import dataclasses
 from dataclasses import dataclass, field
 from datetime import datetime
 from typing import Any, Optional
 from uuid import UUID
 
+import trafaret as t
+
+from ai.backend.common import validators as tx
+from ai.backend.common.types import ClusterMode, SessionTypes
 from ai.backend.manager.data.session.types import SessionStatus
+
+overwritten_param_check = t.Dict({
+    t.Key("template_id"): tx.UUID,
+    t.Key("session_name"): tx.SessionName,
+    t.Key("image", default=None): t.Null | t.String,
+    tx.AliasedKey(["session_type", "sess_type"]): tx.Enum(SessionTypes),
+    t.Key("group", default=None): t.Null | t.String,
+    t.Key("domain", default=None): t.Null | t.String,
+    t.Key("config", default=None): t.Null | t.Mapping(t.String, t.Any),
+    t.Key("tag", default=None): t.Null | t.String,
+    t.Key("enqueue_only", default=False): t.ToBool,
+    t.Key("max_wait_seconds", default=0): t.Int[0:],
+    t.Key("reuse", default=True): t.ToBool,
+    t.Key("startup_command", default=None): t.Null | t.String,
+    t.Key("bootstrap_script", default=None): t.Null | t.String,
+    t.Key("owner_access_key", default=None): t.Null | t.String,
+    tx.AliasedKey(["scaling_group", "scalingGroup"], default=None): t.Null | t.String,
+    tx.AliasedKey(["cluster_size", "clusterSize"], default=None): t.Null | t.Int[1:],
+    tx.AliasedKey(["cluster_mode", "clusterMode"], default="SINGLE_NODE"): tx.Enum(ClusterMode),
+    tx.AliasedKey(["starts_at", "startsAt"], default=None): t.Null | t.String,
+    tx.AliasedKey(["batch_timeout", "batchTimeout"], default=None): t.Null | tx.TimeDuration,
+}).allow_extra("*")
 
 
 @dataclass
